Move back apis, models related to endpoint registry to API module
