+++ components/apimgt/org.wso2.carbon.apimgt.api/src/main/java/org/wso2/carbon/apimgt/api/endpoint/registry/api/EndpointRegistryResourceAlreadyExistsException.java
@@ -15,7 +15,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.carbon.apimgt.impl.endpoint.registry.api;
+package org.wso2.carbon.apimgt.api.endpoint.registry.api;
+
+import org.wso2.carbon.apimgt.api.endpoint.registry.api.EndpointRegistryException;
 
 public class EndpointRegistryResourceAlreadyExistsException extends EndpointRegistryException {
