+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.endpoint.registry/src/gen/java/org/wso2/carbon/apimgt/rest/api/endpoint/registry/SettingsApiService.java
@@ -7,7 +7,7 @@ import org.apache.cxf.jaxrs.ext.MessageContext;
 import org.apache.cxf.jaxrs.ext.multipart.Attachment;
 import org.apache.cxf.jaxrs.ext.multipart.Multipart;
 
-import org.wso2.carbon.apimgt.impl.endpoint.registry.api.EndpointRegistryException;
+import org.wso2.carbon.apimgt.api.endpoint.registry.api.EndpointRegistryException;
 
 
 import java.util.List;
@@ -19,6 +19,6 @@ import javax.ws.rs.core.SecurityContext;
 
 import org.wso2.carbon.apimgt.rest.api.endpoint.registry.RegistriesApi.*;
 
-@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJAXRSCXFCDIServerCodegen", date = "2020-05-28T10:45:00.026+05:30[Asia/Colombo]")public interface SettingsApiService {
+@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJAXRSCXFCDIServerCodegen", date = "2020-05-28T12:57:10.924+05:30[Asia/Colombo]")public interface SettingsApiService {
         public Response getSettings(MessageContext messageContext) throws EndpointRegistryException;
 }
