+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.endpoint.registry/src/main/resources/server-templates/api.mustache
@@ -4,7 +4,7 @@ package {{package}};
 {{/imports}}
 import {{package}}.{{classname}}Service;
 import {{package}}.impl.{{classname}}ServiceImpl;
-import org.wso2.carbon.apimgt.impl.endpoint.registry.api.EndpointRegistryException;
+import org.wso2.carbon.apimgt.api.endpoint.registry.api.EndpointRegistryException;
 
 import org.apache.cxf.jaxrs.ext.MessageContext;
 import org.apache.cxf.jaxrs.ext.multipart.Attachment;
