+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.endpoint.registry/src/main/resources/server-templates/apiService.mustache
@@ -7,7 +7,7 @@ import org.apache.cxf.jaxrs.ext.MessageContext;
 import org.apache.cxf.jaxrs.ext.multipart.Attachment;
 import org.apache.cxf.jaxrs.ext.multipart.Multipart;
 
-import org.wso2.carbon.apimgt.impl.endpoint.registry.api.EndpointRegistryException;
+import org.wso2.carbon.apimgt.api.endpoint.registry.api.EndpointRegistryException;
 
 {{#imports}}import {{import}};
 {{/imports}}
